+++ package.json
@@ -9,6 +9,7 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "framer-motion": "^11.3.29",
     "next": "14.2.5",
     "react": "^18",
     "react-dom": "^18",
